+++ testdata/Arraybool_test.go
@@ -7,6 +7,8 @@ import (
 )
 
 func TestBoolArray(t *testing.T) {
+	// BOOLEAN is not a valid column data type on Db2 for z/OS.
+	SkipOnPlatform(t, PlatformZOS)
 	if BoolArray() != nil {
 		t.Error("Error at BoolArray")
 	}