+++ testdata/Arraygraphic_test.go
@@ -21,7 +21,7 @@ func GraphicArray_1() error {
 
 	db.Query("DROP table " + tableOne)
 
-	_, err := db.Exec("CREATE table " + tableOne + "(col1 int, col2 graphic(5))")
+	_, err := db.Exec("CREATE table " + tableOne + "(col1 int, col2 graphic(5))" + CCSIDUnicodeClause())
 	if err != nil {
 		fmt.Println("Exec error: ", err)
 		return err